backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic[dotenv]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
aiofiles
requests
jinja2 
numpy
python-dotenv
scikit-learn
SpeechRecognition
itsdangerous
pandas
librosa
praat-parselmouth
scipy